# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QROS — Guest, Staff & Admin API
  version: 1.0.0
  summary: Hợp đồng REST cho hệ thống đặt đồ uống qua QR tại bàn.
  description: |
    Hợp đồng này là **nguồn sự thật** cho ba phía triển khai: Spring Boot (server),
    Next.js (web khách + KDS + quản trị), và các job Python theo lịch. Viết tay trước, sinh code sau —
    không sinh spec từ annotation của backend, vì làm vậy thì backend luôn "đúng" theo định nghĩa
    và hợp đồng mất tác dụng.

    Mỗi operation mang `x-requirements` trỏ về mã yêu cầu trong `PRD-QROS-001` v1.1.

    ## Bốn quy tắc bất biến

    1. **Không có trường giá nào đi lên từ client.** Mọi schema request đặt món đều
       `additionalProperties: false` và không khai báo trường giá. Client gửi `unitPrice`
       sẽ bị chặn ngay ở tầng xác thực lược đồ, trả `400 PRICE_NOT_ACCEPTED` — `ADR-06`.
    2. **Ba vùng, ba chuỗi filter độc lập.** `/guest/**` không bao giờ dùng chung
       cấu hình bảo mật với `/staff/**` hay `/admin/**`.
    3. **Tiền là số nguyên đơn vị đồng.** Không dùng số thực ở bất kỳ đâu.
    4. **Mọi lỗi theo RFC 7807**, kèm phần mở rộng `code` (chuỗi ổn định để máy đọc)
       và `traceId`.
  contact:
    name: QROS Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.qros.example.com
    description: Production
  - url: https://staging-api.qros.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local
...
security: []
paths:
...
  /api/v1/guest/menu:
    get:
      operationId: getMenu
      tags: [guest-menu]
      summary: Thực đơn của chi nhánh trong phiên
      description: |
        Trả về toàn bộ thực đơn đang phát hành, kèm trạng thái còn/hết theo thời gian thực.
        Đáp ứng có `ETag`; client gửi `If-None-Match` để nhận `304` — đây là đường đi
        chịu tải lớn nhất hệ thống, xem `NFR-PERF-01`.
      x-requirements: [FR-CUS-03, FR-CUS-04, NFR-PERF-01]
      security: [{ tableSession: [] }]
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string, examples: ['public, max-age=60'] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Menu' }
        '304':
          description: Thực đơn chưa đổi.
        '401': { $ref: '#/components/responses/Problem' }
  /api/v1/guest/menu/items/{itemId}:
    get:
      operationId: getMenuItem
      tags: [guest-menu]
      summary: Chi tiết một món
      description: Trả chi tiết món trong thực đơn của đúng chi nhánh thuộc phiên bàn hiện tại.
      x-requirements: [FR-CUS-03, FR-CUS-05]
      security: [{ tableSession: [] }]
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuItem' }
        '404': { $ref: '#/components/responses/Problem' }

  # ────────────────────────────────────────────────────────────────
  #  ĐẶT MÓN
  # ────────────────────────────────────────────────────────────────
  /api/v1/guest/orders:
    post:
      operationId: placeOrder
      tags: [guest-order]
      summary: Đặt món
      description: |
        **Điểm nhạy cảm nhất của toàn bộ hợp đồng.**

        Client chỉ gửi định danh và số lượng. Máy chủ tính lại toàn bộ giá, phụ phí, thuế
        và giảm giá từ catalog trong cùng một giao dịch. Lược đồ request đặt
        `additionalProperties: false`, nên mọi trường lạ — đặc biệt là trường giá —
        bị từ chối ngay ở tầng xác thực.

        `Idempotency-Key` là **bắt buộc**. Gửi lại cùng khoá trong 24 giờ trả về đúng đơn cũ
        kèm `200` thay vì tạo đơn mới. Đây là lớp bảo vệ chống chạm hai lần và chống thử lại
        sau khi mất mạng.
      x-requirements: [FR-CUS-08, FR-CUS-09, ADR-06, EC-01]
      security: [{ tableSession: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '201':
          description: Đơn đã được tạo.
          headers:
            Location:
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '200':
          description: Khoá idempotency đã dùng — trả lại đơn đã tạo trước đó.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: |
            `PRICE_NOT_ACCEPTED` — client gửi kèm trường giá.
            `VALIDATION_FAILED` — payload sai lược đồ.
          $ref: '#/components/responses/Problem'
        '409':
          description: |
            `PRICE_CHANGED` — giá đổi giữa lúc xem và lúc đặt; phần mở rộng `currentTotal`
            chứa bảng giá mới để client hiển thị xác nhận lại.
            `ITEM_SOLD_OUT` — món vừa hết.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/PriceChangedProblem' }
        '422':
          description: '`IDEMPOTENCY_KEY_REUSED` — khoá đã dùng cho một yêu cầu khác nội dung.'
          $ref: '#/components/responses/Problem'
        '429':
          description: '`ORDER_RATE_LIMITED` — vượt hạn mức 3 đơn / 5 phút / phiên.'
          $ref: '#/components/responses/RateLimited'

    get:
      operationId: listSessionOrders
      tags: [guest-order]
      summary: Các đơn trong phiên hiện tại
      description: Liệt kê các đơn và tổng tiền thuộc đúng phiên bàn hiện tại.
      x-requirements: [FR-CUS-10, FR-CUS-11]
      security: [{ tableSession: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [orders, sessionTotal]
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
                  sessionTotal: { $ref: '#/components/schemas/Money' }
                  paymentStatus:
                    type: string
                    enum: [UNPAID, PENDING_CASH, AUTHORIZING, SETTLED]
  /api/v1/guest/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [guest-order]
      summary: Chi tiết một đơn
      description: |
        Kiểm tra quyền sở hữu ở cấp đối tượng: đơn phải thuộc đúng phiên đang gọi.
        Không đủ điều kiện thì trả `404`, **không** trả `403` — tránh xác nhận sự tồn tại
        của đơn thuộc bàn khác.
      x-requirements: [FR-CUS-10, NFR-SEC-06]
      security: [{ tableSession: [] }]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/guest/orders/{orderId}/cancel:
    post:
      operationId: cancelOrderByGuest
      tags: [guest-order]
      summary: Khách tự huỷ đơn
      description: Chỉ chấp nhận khi đơn còn ở `PENDING`. Từ `PREPARING` trở đi phải qua nhân viên.
      x-requirements: [FR-CUS-08, EC-06]
      security: [{ tableSession: [] }]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Đã huỷ.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '422':
          description: '`INVALID_TRANSITION` — đơn đã rời khỏi `PENDING`.'
          $ref: '#/components/responses/Problem'
  /api/v1/guest/staff-calls:
    post:
      operationId: callStaff
      tags: [guest-order]
      summary: Gọi nhân viên
      description: Giới hạn 1 lần / 90 giây / bàn để chống spam.
      x-requirements: [FR-CUS-12]
      security: [{ tableSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason:
                  type: string
                  enum: [REFILL_WATER, CLEAN_TABLE, PAYMENT_HELP, OTHER]
                note: { type: string, maxLength: 200 }
      responses:
        '202':
          description: Đã chuyển tới nhân viên.
        '429': { $ref: '#/components/responses/RateLimited' }

  # ────────────────────────────────────────────────────────────────
  #  THANH TOÁN
  # ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    tableSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWS `EdDSA`, TTL 90 phút trượt theo hoạt động, gắn với `deviceId`.
        Claim: `iss`, `aud`, `sub`, `exp`, `iat`, `jti`, `sid`, `tid`, `scope`, `tv`.
        Không chứa thông tin cá nhân.
    staffSession:
      type: apiKey
      in: cookie
      name: qros_session
      description: |
        Cookie `HttpOnly; Secure; SameSite=Strict; Path=/api`. TTL 15 phút,
        làm mới qua `/auth/refresh`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 trên raw body, khoá chia sẻ với cổng thanh toán.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUIDv4. Lưu 24 giờ. Gửi lại cùng khoá trả về kết quả cũ.
      schema: { type: string, format: uuid }
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ItemId:
      name: itemId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Lỗi theo RFC 7807.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Vượt hạn mức.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 kèm hai phần mở rộng bắt buộc của QROS.
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: |
            Thông điệp cho người dùng cuối. **Không bao giờ** chứa stack trace,
            câu truy vấn, hay chi tiết nội bộ — `NFR-SEC` A05.
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Mã ổn định để máy đọc, ví dụ `PRICE_NOT_ACCEPTED`.
          examples: [PRICE_NOT_ACCEPTED, ITEM_SOLD_OUT, INVALID_TRANSITION]
        traceId:
          type: string
          description: Tương quan với truy vết OpenTelemetry để nhân viên tra cứu.
    PriceChangedProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            currentTotal: { $ref: '#/components/schemas/Money' }
            changedLines:
              type: array
              items:
                type: object
                properties:
                  menuItemId: { type: string, format: uuid }
                  oldUnitPrice: { $ref: '#/components/schemas/Money' }
                  newUnitPrice: { $ref: '#/components/schemas/Money' }
    Money:
      type: object
      description: |
        Tiền luôn là **số nguyên đơn vị đồng**. Không dùng số thực ở bất kỳ đâu
        trong hệ thống, kể cả khi hiển thị.
      required: [amount, currency]
      properties:
        amount:
          type: integer
          format: int64
          examples: [55000]
        currency:
          type: string
          enum: [VND]
      additionalProperties: false
...
    Menu:
      type: object
      required: [storeId, categories, updatedAt]
      properties:
        storeId: { type: string, format: uuid }
        updatedAt: { type: string, format: date-time }
        categories:
          type: array
          items:
            type: object
            required: [id, name, items]
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              displayOrder: { type: integer }
              items:
                type: array
                items: { $ref: '#/components/schemas/MenuItem' }
    MenuItem:
      type: object
      required: [id, name, basePrice, available, variants]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        imageUrl: { type: string, format: uri }
        basePrice: { $ref: '#/components/schemas/Money' }
        available:
          type: boolean
          description: Cập nhật realtime khi barista báo hết nguyên liệu — `FR-BAR-05`.
        allergens:
          type: array
          description: Nhãn cảnh báo dị ứng. Sai ở đây là rủi ro sức khoẻ, không phải lỗi hiển thị.
          items:
            type: string
            enum: [MILK, PEANUT, GLUTEN, SOY, EGG, NUTS]
        attributes:
          type: array
          items:
            type: string
            enum: [DAIRY_FREE, CAFFEINE_FREE, LOW_SUGAR, VEGAN, HOT, COLD]
        variants:
          type: array
          minItems: 1
          items:
            type: object
            required: [id, name, price]
            properties:
              id: { type: string, format: uuid }
              name: { type: string, examples: ['Size L'] }
              price: { $ref: '#/components/schemas/Money' }
              available: { type: boolean }
        optionGroups:
          type: array
          items: { $ref: '#/components/schemas/OptionGroup' }
    OptionGroup:
      type: object
      required: [id, name, selection, options]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, examples: ['Mức đường'] }
        selection:
          type: string
          enum: [SINGLE, MULTIPLE]
        required: { type: boolean }
        minSelect: { type: integer, minimum: 0 }
        maxSelect: { type: integer, minimum: 1 }
        options:
          type: array
          items:
            type: object
            required: [id, name, surcharge]
            properties:
              id: { type: string, format: uuid }
              name: { type: string, examples: ['50%'] }
              surcharge: { $ref: '#/components/schemas/Money' }
              available: { type: boolean }
    CreateOrderRequest:
      type: object
      additionalProperties: false
      description: |
        ⚠ `additionalProperties: false` ở đây và ở `CreateOrderLine` là biện pháp cưỡng chế
        `ADR-06` tại tầng hợp đồng. Không được nới lỏng trong bất kỳ hoàn cảnh nào.
      required: [lines]
      properties:
        lines:
          type: array
          minItems: 1
          maxItems: 8
          description: Tối đa 8 món mỗi đơn — hạn mức chống lạm dụng, lớp 7 mục 5.3.2.
          items: { $ref: '#/components/schemas/CreateOrderLine' }
        note: { type: string, maxLength: 200 }
    CreateOrderLine:
      type: object
      additionalProperties: false
      required: [menuItemId, variantId, quantity]
      properties:
        menuItemId: { type: string, format: uuid }
        variantId: { type: string, format: uuid }
        optionIds:
          type: array
          maxItems: 10
          items: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 20 }
        note:
          type: string
          maxLength: 200
          description: |
            Văn bản tự do do khách nhập. Được làm sạch trước khi hiển thị trên KDS,
            và coi là **dữ liệu không tin cậy** nếu lọt vào ngữ cảnh trợ lý — `NFR-SEC-16`.
        addedBy: { type: string, maxLength: 24 }
    Order:
      type: object
      required: [id, status, lines, total, placedAt, version]
      properties:
        id: { type: string, format: uuid }
        shortCode:
          type: string
          description: Mã ngắn hiển thị cho khách và barista.
          examples: ['A04-17']
        status:
          type: string
          enum: [PENDING, CONFIRMED, PREPARING, READY, SERVED, COMPLETED, CANCELLED, EXPIRED]
        lines:
          type: array
          items: { $ref: '#/components/schemas/OrderLine' }
        subtotal: { $ref: '#/components/schemas/Money' }
        discount: { $ref: '#/components/schemas/Money' }
        total: { $ref: '#/components/schemas/Money' }
        placedAt: { type: string, format: date-time }
        estimatedReadyAt:
          type: string
          format: date-time
          description: Tính theo độ dài hàng đợi hiện tại — `FR-CUS-10`.
        requiresStaffConfirmation:
          type: boolean
          description: '`true` cho đơn đầu tiên của phiên lạ — cổng xác nhận ở `EC-03`.'
        version: { type: integer }
    OrderLine:
      type: object
      required: [id, menuItemId, name, quantity, lineTotal, status, version]
      properties:
        id: { type: string, format: uuid }
        menuItemId: { type: string, format: uuid }
        name: { type: string }
        variantName: { type: string }
        optionNames:
          type: array
          items: { type: string }
          examples: [['Đường 50%', 'Ít đá', 'Thêm trân châu']]
        quantity: { type: integer }
        unitPrice:
          $ref: '#/components/schemas/Money'
          description: Chỉ có ở **đáp ứng**. Không bao giờ chấp nhận ở request.
        lineTotal: { $ref: '#/components/schemas/Money' }
        note: { type: string }
        addedBy: { type: string }
        status:
          type: string
          enum: [PENDING, CONFIRMED, PREPARING, READY, SERVED, CANCELLED]
        version:
          type: integer
          description: Dùng cho khoá lạc quan qua `If-Match` — `EC-06`.
...